requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
schwabdev